LayoutTests/js/dfg-switch-imm-negative-expected.txt

Tests that a switch statement with a negative integer doesn't cause weirdness.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS foo(-1) is "foo"
PASS foo(0) is "bar"
PASS foo(1) is "baz"
PASS successfullyParsed is true

TEST COMPLETE
